"3:45pm. https://agda.github.io/agda-stdlib/README.Data.AVL.html
Only a little bit let in the video is left, but I found this readme a while ago. I'll go through it once I am done. It seems it might be possible to use maps for my purpose.

4:10pm. Done with the talk. In contrast to yesterday, I really did crap today. But issues like this always act as a drain on my time.

https://agda.github.io/agda-stdlib/README.Data.AVL.html

Let me play with this for a while. I guess I'll go with AVL trees after all since they can be used to prove things.

4:25pm.

```
module _ {v} {V : Value v} where

  private
    Val = Value.family V

  empty : Tree V
  empty = tree $′ Indexed.empty ⊥⁺<⊤⁺

  singleton : (k : Key) → Val k → Tree V
  singleton k v = tree (Indexed.singleton k v ⊥⁺<[ k ]<⊤⁺)

  insert : (k : Key) → Val k → Tree V → Tree V
  insert k v (tree t) = tree $′ proj₂ $ Indexed.insert k v t ⊥⁺<[ k ]<⊤⁺

  insertWith : (k : Key) → (Maybe (Val k) → Val k) →
               Tree V → Tree V
  insertWith k f (tree t) = tree $′ proj₂ $ Indexed.insertWith k f t ⊥⁺<[ k ]<⊤⁺

  delete : Key → Tree V → Tree V
  delete k (tree t) = tree $′ proj₂ $ Indexed.delete k t ⊥⁺<[ k ]<⊤⁺
```

I need to push through this. It has been a while since I've been this intimidated, but I should be able to cut through.

4:40pm. This is so troublesome. Somehow at this point, having that map is even worse than not having it. It presents me with a choice of picking either having to learn how to use the damn map, or alternatively implementing an objectively inferior solution.

It is like that teleporter paradox from game theory.

4:45pm. I am trapped right now.

What I am going to do tomorrow is try to rebuild the AVL tree on my own, mostly by copy pasting the relevant segments top down and tracing through them. I am putting proving things about CFR on hold.

4:50pm. I should have started off today doing that instead of hemming and hawing about whether to learn this at all. In my desire to avoid wasting time, I wasted even more. Things turn out like that.

I am going to give AVL trees my full focus tomorrow and for the next few days or so.

It is a reasonably complex piece of Agda code and if I can understand everything that is in there, I will have made a remarkable increase in my understanding of the language. Hopefully, having the maps paddled will be enough to fill my data structure needs for the days to come.

4:50pm. Yeah, no matter how you look at it, maps are important. I spent a few days rewriting the thing in Lean too only to mess it up in the end. Agda thankfully does in fact do trichonomous comparison, so I won't have to deal with that idiocy of not being able to prove equality of the keys.

The code in the library is fairly elaborate, mostly because my grasp on Agda is not that strong yet. Still there is nothing in there that I cannot figure out with some effort.

I just need to make up my mind to do it. Once an arrow is fired in the right direction, it will hit the target.

5pm. I'll close Emacs here as I won't be doing any more programming for the day, but rather than lazing, let me watch a lecture by Wildberger.

5:05pm. If I want to use Agda then I am going to have to make up my mind to get deeply familiar with the stand library. It is not like there is much of it. Probably something like AVL trees is the most complex piece of it. Well, I guess that reflection using solvers would be worse, but nevermind that for now.

Probably at this point, my future trajectory has been laid out for me.

Due to circumstances, I've been placed solidly in the Agda camp which is on the dependent PL side as opposed to the proof assistant side like Coq and Lean. I am perfectly fine with that actually. This is the side I've wanted to be on from the start.

I've admitted my weakness - I cannot judge the potential utility of tactics that proof assistants rely on, but I sure can judge beauty and elegance of languages. Agda definitely gets top marks here.

In fact, I'd give it the top mark. It does so much right.

It really is a pity none of the dependently typed languages can be used to do real world work at the moment.

5:10pm. But I am sure that one day that will change.

It is also a pity that Lean fell through for me. But I gave it a chance. It can only have itself to blame.

5:15pm. It is time for me to start getting serious, not so much as a programmer, but as scientist. I gave it my best with Spiral. Not being able to do it essentially wasted my entire year.

From here up to 2020, the next two months, I will be getting familiar with Agda while doing some simple proofs, and then from there I am going to be a total grinder. I will stay in this seat that I am perennially glued to and churn out those proofs. I will formalize a lot of the field of RL until I have a good map of it.

I will dedicate myself wholesomely to logic.

5:20pm. I actually had not realized up to now that the destiny of every dedicated programmer is to become a logician in the end. Some food for thought."