"8:20am. I went to bed at 6:30pm due to thunder...and it was a relief as I am stressed. I just did not feel like chilling at that time. The highlight of today's night seems to be me thinking more seriously of how to put dependent types into Spiral. It really might be possible to fuse the passes together so that top-down typechecking is done during the partial evaluation pass.
8:25am. I guess now I finally have enough experience in Agda and such to suss out the design for an efficient DT language.

Nonetheless, forget that. What I need to do is appreciate what Agda has. Its pure type system is tailor made for proving things and that is what I should do. Spiral with DTs would be an impure language and that sort of goodness would be lost in it. It would be more of an implementation language with typecase and all.

At this time I do not think a DT with that feature exists.

I had a lot of time to think last night. I've been looking down on DT langs too much. Sure they might not be efficient, and the kinds of programs they can prove properties about is limited. But what I have no appreciated till now is that proving is a miracle. It took so much effort in PLs just to get to the point where that is possible.

8:35am. I just saw them as weird. It is really, really unfortunate that impurity is so hard to reason about. If it weren't things would have gone so much differently. Damn this shitty world. Really, if something can go wrong, it will won't it?

At any rate, I need to prove things. I absolutely need to attain the highest skill possible in this.

8:40am. So towards that end, I must study AVL trees. Where would I be without them?

All this sleep has gotten me tired. Let me chill a bit and then I will start."