"12:50pm. Done with breakfast. Shinchou and then I will start programming for real.
1:15pm. "I have two lives. I may have lost one, but I still have one left."

Holy shit, so that is why he kept blasting the remains.

1:20pm. It turns out the episode last week was not a recap. Let me go for the last one. I can't possibly stop here.

1:45pm. That was really satisfying. Ok, chores and then I start.

2:05pm. It is time to start.

2:10pm. No seriously. I've decided, I am not going to be switching repos. I only wanted to do this because I did not want to spend time cleaning up the VS project stuff, but there is no need for that now.

Let me just open it in the new VS.

Oh it seems there is no need to mess with the project files. It opens as is. Great.

Let me update the IDE and then I will think of how to start. Damn, it really is 1Gb. For once the installed did not lie.

2:20pm. I am thinking.

This redesign is really a shitload of work. A lot of what I am doing will be novel as well.

The first order of business will be to refamiliarize myself with the old codebase.

This won't take too long. My code is pretty good, so it should come back to me easily enough despite being 4.5k lines of code.

2:40pm. Updated the readme a bit.

With this I am finally home. Spiral is fundamentally what I should be working on. Enough math. Enough ML.

I've figured out how to resolve all the compilation issues. And I figured out how to make the language ergonomic.

This time I will really finish the language. I will make Spiral into what it really should be, consequences to myself be damned.

Spiral is my heart. Spiral is my home."