"1:50pm. Ok, let me chill a bit more, do the chores and then I will resume by asking some questions about .NET references on SO.
1:55pm. Let me start. Since I want to get the answer ASAP, let me do that research and then I'll do the chores.

https://stackoverflow.com/questions/750947/net-unique-object-identifier

"The reference is the unique identifier for the object. I don't know of any way of converting this into anything like a string etc. The value of the reference will change during compaction (as you've seen), but every previous value A will be changed to value B, so as far as safe code is concerned it's still a unique ID."

Ah, I see. So then I definitely can't use the comparison. Still if that is true, how exactly do the reference dictionaries work...

Ah...the OP mentioned that two instances can get the same hash. That means that an object gets a hash assigned to it at random during creation or something like that.

https://docs.microsoft.com/en-us/dotnet/api/system.runtime.serialization.objectidgenerator?redirectedfrom=MSDN&view=netframework-4.8

This class might be good to keep in mind if I need explicit IDs which I probably won't.

But it does not say whether this method uses weak tables. It probably does not. So I should avoid it.

2:05pm. I got it.

https://docs.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.runtimehelpers.gethashcode?redirectedfrom=MSDN&view=netframework-4.8#overloads

`RuntimeHelpers.GetHashCode` together with references equality is what I can use to implement...

Ah no wait shit. What do I do about the situation where references are not equal, but the hashcodes aren't. That will a problem. I can use this after all.

It might be good to ask how it would be possible to implement comparison just by reference.

2:25pm. https://docs.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.conditionalweaktable-2?view=netframework-4.8

The CWT does not lock in data.

Ok...

2:30pm. I need to rethink hash consing now. I cannot really do that as I have up to now. I am going to have to implement my own equality for TypedData - this won't be hard as I just need to try reference equality first while checking for cycles.

But what about hashing? If I have to do marking, should I just return zero when I reach a place that has been marked?

2:35pm. Ogh, shit.

Sure I can do recursive blocks easily enough using the idea I had yesterday, but that will give me significant difficulties in other places.

2:45pm. I have an idea. What I am going to do is wrap all the cyclical blocks to mark them as such. Their hash will be 0 and they will use reference equality.

I'll make a little wrapper class and use it on `TyFun` in recursive blocks.

That will do the trick.

2:50pm. It is actually pretty troublesome for me to understand how to deal with recursive blocks in the usual manner while at the same manner do variable filtering which I need to do for join points. It is straightforward with standard recursive functions, but blocks are a different beasts entirely.

Hmmm...actually...

2:55pm. Let me do the chores here. All this recursive thought it making my mind diverge."