"10:05am. Wow the 20s are here. Will I be able to accomplish anything in this decade, I wonder? Whatever it turns out to be, I will do my best.
Right now the decision of whether or not to put in higher rankeds types into Spiral is really torturing me. Even if I do not put them in, it is not like everything suddenly becomes easy.

I have no idea. I just need more time to think about this.

10:15am. Agh, this will take some time. Let me update Simulacrum page.

10:45am. Let me leave a comment as well. I was a bit too lazy this year.

12pm. "Let me say a few extra words as I was lazy this year and will make the above my monthly PL thread review as well. I made this decision because the review itself captures all that I wanted to say, but this place is special to me, so I want do a bit more here.

Let's see...2022 is just too early as the date for the Singularity. The memristors fell through, the hardware is not there, the algorithms are not there. They might be here in 2029 and should be, but I can't see it happening in 2022. If does happen at that time, I am certain it won't be due to anything I personally did. My ambition for the next two years is to make v0.2 of Spiral and given how long the previous versions took, that seems realistic for how long it will take me to wrap the language up.

It is not like it necessarily needs to take 2 years, things like a simple type systems and editor support could certainly be done much more quickly than that. But also, it is not like making a partial evaluator should take that long either, and yet it did. But after I got the requisite experience, I found that I could do it in a month - that is what my [February 2019 rewrite](https://github.com/mrakgr/The-Spiral-Language/tree/v0.1) of Spiral demonstrated.

Back in 2017 I could have only dreamed about such an ability. Whenever I try to skip steps, I stumble. I always want to grasp beyond my reach, but I've never even once managed to break the laws of causality. So v0.2 is going to need its investment of time.

But once I have it; once I have the design, the implementation and the editor support, I will never have to look at another language again. v0.2 will be pretty plain in terms of type system features. It certainly won't be anything to write about compared to dependently typed languages, or even what Haskell is today. Nonetheless what it will have are those generative constraints I've talked about.

Because of them Spiral v0.2 will have by far the best generative capabilities of any statically typed language in existence. The whole issue of GPU backends and ML libraries will just fall into the background once v0.2 is on the scene, and I will be able to program in comfort knowing that the work is done. I already solved it halfway with the previous version and now it is the time to finish the job.

No doubt, the Inspired will dream up designs and create languages that Spiral v0.2 cannot even compare in power and usability, but nonetheless I am confident that the idea of combining constraint solving with partial evaluation is going to go down as one of the greatest ideas in PL design ever. It just solves some very difficult problems with static typing quite cleanly that even dependent types don't.

Even those great languages of the future might end up doing it. So v0.2 needs to happen, and nobody but me can do it. There are plenty of other things I could do, but none of them are as good as this. Spiral has to be done at some point and that is when my focus will be freed to actually do the things I said I would.

But what about money? I've long been trying to angle it without success so that my increases in programming skill correspond to increases in financial capacity, but this world is not that sort of game apparently. It might get closer to that once AI stops being a toy and programming skill becomes a closed loop leading to further increases in that skill.

I really do envy the Inspired quite a bit. Imagine if I could do everything I did in the past 5 years in something like 5 days. That should be easily doable for them. The future is just so unfair. My efforts take up the entirety of my life and yet they are so light compared to what they could be. My limitations are so burdensome and onerous. A stroke of bad luck could happen and I will not be there to see the Singularity happen. I just have no idea what is the point of even being a human. Why must this era exist? When those of the future look back on this, what do they really see? Is it a nightmare...or the first flame?

If they need a reminder of what the purpose of life is, then let them find this post and be reminded that the main axiom is power.

As pathetic as it is to hear something like that from a human, I will never give up on this desire or this ambition. I am just a tiny plant in a big forest looking upwards towards the sky and thinking: 'Wouldn't it be great if those assholes stopped blocking the shade.'

The fire will happen and the seeds I've planted will sprout."

12:05pm. Let me go with this. This is good enough.

12:15pm. I guess I'll start slacking here.

Even if I decide to kick out HRTs, I still have HKTs to deal with. I am just not really sure how to deal with those. Right now I am thinking of how I would do things like printing and such on the partial evaluation side.

I am definitely missing something.

Just for the sake of HKTs, do I not need foralls and type variables at the type level?

Something slipped my mind. Do HKTs not need to be type level functions? Sure I could make them similar to inlineables, but shouldn't I be better than that?

There is literally no need for them to be macros in their composition...

Or no wait. Maybe there is. If I make type level functions foralls as opposed to inlineables, how will I deal with `real nominal` types?

12:20pm. Wait, actually, if I do make macro like type level functions, could I not do that way the HRTs like I originally intended.

Like if you ever try to box `<type function>` it would be quite obvious.

12:25pm. No, HKTs are still quite different from HRTs. For HRTs I would need explicit foralls on the type level.

During partial evaluation I really only want to deal with inlineables of various stripes and monomorphic types.

Yeah, I can't really go beyond first order logic unless I want to put in some quite significant effort into it.

12:30pm. Which I do not really want.

But HKTs will not give me much trouble. I will be able to deal with them quite fine if I make them inlineables.

This is the way to go.

Let me take a break here."