"2:50pm. I think I will just step away from the screen for a while like I said I would yesterday. It is thinking time. Last month I spent quite a value thinking about the design for the language.
But now, I really do need to think about what would eventually use it for.

Clearly I am starting to develop feelings for mathematics.

I just do not know how to express them in a positive manner.

After the experiences of the last year, I am not really going to be able to just abandon mathematics. But I am not going to be able to just go through the ML papers like I wanted to either.

Really, as you go from infinitism to programming, the proofs do get harder, but it is not like you get nothing in return for exiling yourself from paradise.

[Randomized testing.](https://fscheck.github.io/FsCheck/)

When I did that first CFR proof, I could really feel the power of this.

At all really rests on these lightweight methods. I really do think that regular theorem proving should be kept for special things. I wish it were otherwise, but DTLs really are quite hard deal with.

Randomized testing is rarely talked about, but it really might be the killer application that finitists needed, but did not quite have in the early 20th century, along with computers.

3pm. Also, I just realized that I never even pushed the reply to the guy. I must be even madder than I thought to miss this. Still, I think I will just leave it at that since the flow is going against me anyway.

I am not really the kind of person to fish for negative votes.

Even if I made the post it is not like it would have added anything new to the conversation. That guy is just acting like a damage controller for the consensus. Calling constructive mathematics a flavor (therefore normalizing the deviance), and saying rejecting classical logic is unreasonable, I can only spit on this.

3:10pm. I was really fucking tooled like nothing else by classical mathematicians. I really need to think of a suitable reply to that, but whatever that reply is it cannot take the shape of arguing on the Internet.

There should be something better I can do to pay respects to my new principles.

3:15pm. Rejecting faulty assumptions is not unreasonable at all. It will really pare down the search space."