"10:45am. I've stewed in my own hatred and misery the whole night, but it is not enough. I need some more.
As much animosity as I've gathered towards the mathematicians, what exactly am I going to do here?

I put in the utmost elite amount of effort into theorem proving.

And yet I cannot reduce the CFR to anything more than it already is. It is a complete failure.

All the proofs I can do are meaningless to me, but the one thing I want to do I cannot.

Maybe it is time to accept it - as one goes from integers to floats, and as one goes from tabular to distributed representations, the ability to say (let alone prove) anything at all using non-empirical methods because sharply difficult, way beyond my ability.

And I certainly cannot believe that a bunch of people who cannot even do numbers properly can be expected to produce great insight about AI. I would be an absolute moron to think that.

At this point I am convinced that math is one giant middle finger towards programming. I wrote that as much in the past months.

But I've sort of been coping and looking for a way out of this nightmare.

Forget it. It cannot be done.

Also, while mathematicians are trash, the fact that they opted for this course of action plainly tells anybody just how difficult rationals really are. And the proofs involving them will only get more difficult once one switches to truncated rational equivalents that are floats.

10:55am. The real world is harsh! Hard!

CFR is half a page long and yet I cannot take it as anything than a primitive!

I am nothing more than mouse wondering what the eagle could see!

...

I've been obsessed all this time in finding a way to make float-using algorithms as easy to reason as the discrete ones, but I am dying here.

As a first step towards my rehabilitation, maybe I need to stop and say: 'It is not just you. This really is hard.'.

The nature of self improvement is one thing, but I saw it as my duty to fight and get anything I can before embracing that darkness. It is no use. My power right now is just too little to cut through this knot.

CFR is not even close to deep learning in difficulty, and yet I cannot do it.

11am. Ahhhh, my dreams...that guy was really mad at me for my comments, but I am the one suffering the most here. I remember in April look at CFR and thinking how great it would be to have the ability to prove things about it. I thought I would get something from going through Software Foundations, but the only thing I learned there was how to do proofs about things like the ones in the book. Trivial things I was not interested in the first place.

11:05am. Maybe I was wrong. Maybe the current behavior of the ML research community is the one that is right. Maybe the randomized testing that they are subconsciously doing is the way science should be conducted.

It is far better to do real science no matter how ad-hoc it is than fake math.

11:15am. Though I have the heart of a villain, my principles do not allow me to reject reality.

If hacking things together is the best that can be done pre-Singularity, then my aim should be to be the best in the world at it.

...

Self-improvement should be done organically.

If math is to be resurrected to its former level of rigor, it needs to happen naturally with steady advances in programming languages.

If I do Spiral v0.2, I will have cleaned up the lower end, leaving only dependent types left as the only real challenge in PL design.

11:20am. I honestly do not think I will be getting to writing DT langs as a human - Spiral has taken up enough of my time as it is, but now that I've come to this point let me just indulge myself a little before breakfast.

I want to check out F* again. I am interested how it deals with things like arrays. Agda and Coq do not even have those constructs, so I am really curious about this particular aspect right now.

Last time I looked at F* I understood nothing of it, but now that I understand type theory I might as well take advantage of this good fortune.

I do not think I will be programming today. I want more time to just stew and form my motivation properly. I still do not have it. I need a concrete goal to which to aim for. There should be one somewhere. Just how will I be the best in the world at ad-hoc science? What would serve as a proof of that?"