"12:35pm. Let me do the chores here. I'll be helping around the house for a while.
I needed an excuse to get away from the screen anyway.

4:30pm. I am back. Somehow the manual labor felt right at this time. I am not in the right mental state to really resume work on Spiral.

https://www.reddit.com/r/MachineLearning/comments/ejbwvb/r_single_biological_neuron_can_compute_xor/

Maybe I should apologize to the ML crowd. The information value of various ML experiments is low, but it might be possible to get somewhere by scaling it in a principled fashion.

https://venturebeat.com/2020/01/02/top-minds-in-machine-learning-predict-where-ai-is-going-in-2020/

It has really been a while since I did anything ML related.

yosuf-bengio: "I believe 2020 will be a great the year for AI hardware.

* First AI accelerator startups will launch their products, thus increasing competition

* Google release their 4th-gen TPUs and already announced their next-gen edge-TPU

* Nvidia introduce their 7nm GPUs

* Intel announces their first PCIe GPU lineup, which will most likely support DL applications

* Advances in compiler technology, e.g. Glow, XLA, MLIR, make it easier for hardware companies to support for common DL frameworks

* Developments in pretraining/self-supervised learning methods will increase the demand for compute"

This post is hope to me.

4:50pm. It seems the new Madoka anime is not out yet.

Well, whatever.

Before I start chilling for real, let me just put in a few words. I need to rant until it runs out.

I need to go back to my old plan.

2019 was sort of a waste - I succeeded in getting good at Agda/etc, but failed at all the reasons taking it up in the first place.

I need to get back to what I was doing in 2018, except this time less maniacal berserking and more measured steps.

5pm. The fundamentals of the whole situation have not changed. I think given how things were, nobody could begrudge me for seriously trying formal theorem proving, but that has been tried and done with.

As a human, I can really be the best at one thing - and since I am fishing for the Singularity, I might as well become the best language creator in the world. That should be a stable existence. And it is an achievable goal. I have the design for it, I just need to create Spiral v0.2.

It is important that I get into the swing of things.

Right now it does feel like the parser is swallowing me up, but the Real Deal would not have much issue with it. If it was me from February 2019, I would have taken care of the parser in a few days and have gone on my merry day.

5:10pm. The best, huh?

If it was some superhuman agent, I bet he would manage to create a language with all the benefits of Spiral and a fully dependent type system. And not only that, he would make the difficult things in DTLs of today easy in his own language.

5:30pm. I have to gather the will to get the ball rolling in a big way.

Basically, rather than redoing the libraries like I did in in v0.1, I want to resume where I stopped in v0.09 while the taking the best aspects of v0.1 (which are mostly under the hood).

I have to start making small gains once again. They are what will get me going.

5:35pm. I was too egoistic to think that I could get anywhere with math. If it was that easy the math guys would have already done it.

But at the same time, I cannot be timid.

Computers are in my blood. I need to get back at looking into ways of upping my programming skill once more.

I do not really feel like competing in the ML library area with Facebook and Google anymore, but nonetheless with Spiral there should be ways of taking advantage of the circumstances in ways that are impossible in other languages.

I need to aim for games.

As much as I'd like to give up poker and do something more interesting, it is still a good target. I need to make a game and throw everything there is at it.

Only then will my soul be at peace.

Mastering games was something that humans were destined for from the start.

And in this era, it should be my calling.

5:40pm. The low style is the most powerful. Randomized testing thing is an important part of its answer to the question of the future.

Randomized testing is the divine technique behind everything. The silent conqueror and the faithful companion.

It was my intention to take risks with my identity for the sake of attaining the power of self-improvement.

If I am a true seeker of the Singularity, I need to start behaving like one. I need to start taking risks with the algorithms themselves.

As a part of that it will be my honor to create Spiral - the language closest to the true one that is computability.

What I have in v0.2 is just the first half of it.

It is a great half, but am I really going to let work be undone, when completing it should vastly improve its productivity and scaling properties?

5:50pm. I said poker, but for v0.2 I really want to try something new. Something other that a ML library or RL agent for poker specifically.

I want a refresher.

Trying something novel will reinvigorate me.

I do not want to build everything from scratch. I want some easy wins this time.

Hmmm...the first thing that comes to mind is programming contests. Maybe I should try to win one using Spiral?

I am really bereft of ideas that don't involve AI in some form.

5:55pm. I want to take a step back from targeting AI directly, and want to work on making a net. I've killed myself in the past 5 years from work and I do not want to do that again. The language itself should be my primary focus, but as a safeguard for the future, not as a primary vehicle towards achieving some aim.

6pm. I kind of want those wins badly, but a sign of maturity is to just do things for their own sake.

A poker agent and a ML library is not that difficult anyway.

What is hard is compressing all of that to nothing.

A sign of skill is to make it all look easy.

...

Here is what is going to happen. I am going to give myself a bit more to think about it. A bit more to rest. And then I am going to take the reigns and take off without looking back.

Nearly 5 years have elapsed since this quest started back in 2015.

By this point, I've pretty much covered the entire spectrum of programming.

There isn't really much for me to explore anymore.

DTLs were probably the hardest challenge I could face, and blew those things out of the water.

I think it is time to settle.

It will take me 1-2 months to get the language up to par with v0.1, and then I will be able to program in it while the work on the typechecker starts in the background.

In six months I will be pretty much done and will be able to look into things like caching compilation residuals and editor support.

6:05pm. This seems like an extreme schedule, but I should be able to meet it.

It is not like in 2017.

I am good now.

And I should be able to prove that at least.

I should be able to prove that I am the appex of humanity in the small area that is programming which is all about composing small informal proofs in one's head.

I do not want to end up being some poser like those rationalists.

I want to earn my own Singularity.

And I have the power to do it.

6:10pm. It is not programming itself that is fun.

It is getting better at it."