"12:35pm. Done with breakfast. Let me do the chores here, and what I will do is just go to bed for a while. I need to think about what my vision for the future should be.
I no longer have a clear cut goal like making a ML library or an agent in it.

So I need to draw that vision in and completely internalize it.

Then I can start coding for real.

1:50pm. Done with chores. Let me just stew about it away from the screen for the next few hours.

I need to do this just for a while longer and then it will be done. I am prepared to do this for as long is necessary till I form the desire in my mind."