"1:15pm. I think I will just rest for a while.
For the past 5 years I've been a fucking robot, and this is the first real crisis that I've ever had since that time.

You can't just recover from having your notion of understanding broken just like that.

Maybe I do have a bit of a mathematician in me. As a programmer I am always on the lookout for novel patterns, algorithms and techniques. Those to me are real gains. This is why I could not stand to abandon functional programming. This is why I could not bare to code GPU kernels in C. That is why I wanted Spiral. It was for the sake of the skills that I had nurtured and not for its own sake.

A code monkey would not have taken this trajectory.

1:20pm. Before, I do anything else, I need to come to an understanding of understanding.

Before I can make gains, I need to make sure that I understand what they are first.

Just imagine it for a while. Make friends with it.

I am going to allow myself a vacation just this once.

For the rest of the week I am not going to hold myself obligated to work. I'll try to make it less than a week as I do not want to waste time, but during this time I am just going to submerge myself in these thoughts.

1:25pm. Fundamentally, understanding is what you interact with. I use all of my imagination to comprehend programs.

One day, a superhuman agent might be able to comprehend things in terms of what humans cannot even imagine.

...

Now, since it is Monday, let me check out Tog. Satanophany might be out as well. And I still haven't watched the new Meguca. Also it is really taking MangaDex a while to come back. It is not dead is it?"