"5:45pm. https://github.com/tensorflow/swift/blob/master/docs/WhySwiftForTensorFlow.md
Usually at this point I would be ending the day, but here I am just continuing from where I left off earlier.

This document is interesting. Right now I am just reading random things.

I think I will just resume tomorrow. I am still depressed, but the fire is starting to burn inside of me.

Since I do not have anything better to do, I should aim to become the best there ever was. I am not going to achieve that by sitting here doing nothing.

I want to have fun building the new language and compiler. I want to find new patterns and techniques like I had done earlier with impure memoization.

5:50pm. I think that this time I am just going to skip the GPU backend as GPUs are absolute shit for RL. They are beyond useless, worse than CPUs.

Something else will turn up, and hopefully by that point I will have gotten some donations to be able to afford an upgrade to my rig.

"Build times could be improved: Developers with large applications (e.g. over 100K lines of code) are not happy with build times. This continues to be a focus of the Swift development community and is expected to improve over time, but this shouldn’t be a problem given the comparatively small size of machine learning models."

Interesting that this is a problem in Swift.

6:15pm. https://github.com/apple/swift/blob/master/docs/TypeChecker.rst
Swift Type Checker Design and Implementation

"Swift's type inference allows type information to flow in two directions. As in most mainstream languages, type information can flow from the leaves of the expression tree (e.g., the expression 'pi', which refers to a double) up to the root (the type of the variable 'three'). However, Swift also allows type information to flow from the context (e.g., the fixed type of the variable 'eFloat') at the root of the expression tree down to the leaves (the type of the numeric literal 2.71828). This bi-directional type inference is common in languages that use ML-like type systems, but is not present in mainstream languages like C++, Java, C#, or Objective-C."

So it is doing the same thing as Haskell and F#. Actual bidirectional type inference cannot fill in implicit vars unfortunately."